+++ mayan/apps/common/icons.py
@@ -2,12 +2,6 @@ from mayan.apps.appearance.classes import Icon
 
 icon_about = Icon(driver_name='fontawesome', symbol='info')
 icon_book = Icon(driver_name='fontawesome', symbol='book')
-icon_current_user_locale_profile_details = Icon(
-    driver_name='fontawesome', symbol='globe'
-)
-icon_current_user_locale_profile_edit = Icon(
-    driver_name='fontawesome', symbol='globe'
-)
 icon_documentation = Icon(driver_name='fontawesome', symbol='book')
 icon_forum = Icon(
     driver_name='fontawesome', symbol='life-ring'