+++ mayan/apps/common/migrations/0016_delete_shareduploadedfile.py
@@ -2,7 +2,6 @@ from django.db import migrations
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ('common', '0015_auto_20200501_0631'),
         ('storage', '0002_auto_20200528_0826')