+++ mayan/apps/common/migrations/0018_delete_userlocaleprofile.py
@@ -3,11 +3,11 @@ from django.db import migrations
 
 class Migration(migrations.Migration):
     dependencies = [
-        ('common', '0016_delete_shareduploadedfile'),
+        ('common', '0017_delete_errorlogentry'),
     ]
 
     operations = [
         migrations.DeleteModel(
-            name='ErrorLogEntry',
+            name='UserLocaleProfile',
         ),
     ]