+++ mayan/apps/common/views.py
@@ -1,24 +1,17 @@
-from django.conf import settings
 from django.contrib import messages
 from django.templatetags.static import static
-from django.urls import reverse_lazy
-from django.utils import timezone, translation
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic import RedirectView
 
 from stronghold.views import StrongholdPublicMixin
 
-from mayan.apps.views.generics import (
-    ConfirmView, SingleObjectEditView, SimpleView
-)
+from mayan.apps.views.generics import ConfirmView, SimpleView
 from mayan.apps.views.mixins import (
     ExternalContentTypeObjectMixin, ObjectNameMixin
 )
 
 from .classes import ModelCopy
-from .forms import (
-    LicenseForm, LocaleProfileForm, LocaleProfileForm_view,
-)
+from .forms import LicenseForm
 from .icons import icon_setup
 from .menus import menu_tools, menu_setup
 from .permissions import permission_object_copy
@@ -28,55 +21,6 @@ from .settings import setting_home_view
 class AboutView(SimpleView):
     extra_context = {'title': _('About')}
     template_name = 'appearance/about.html'
-
-
-class CurrentUserLocaleProfileDetailsView(SimpleView):
-    template_name = 'appearance/generic_form.html'
-
-    def get_extra_context(self, **kwargs):
-        return {
-            'form': LocaleProfileForm_view(
-                instance=self.request.user.locale_profile
-            ),
-            'read_only': True,
-            'title': _('Current user locale profile details'),
-        }
-
-
-class CurrentUserLocaleProfileEditView(SingleObjectEditView):
-    extra_context = {
-        'title': _('Edit current user locale profile details')
-    }
-    form_class = LocaleProfileForm
-    post_action_redirect = reverse_lazy(
-        viewname='common:current_user_locale_profile_details'
-    )
-
-    def form_valid(self, form):
-        form.save()
-
-        timezone.activate(timezone=form.cleaned_data['timezone'])
-        translation.activate(language=form.cleaned_data['language'])
-
-        if hasattr(self.request, 'session'):
-            self.request.session[
-                translation.LANGUAGE_SESSION_KEY
-            ] = form.cleaned_data['language']
-            self.request.session[
-                settings.TIMEZONE_SESSION_KEY
-            ] = form.cleaned_data['timezone']
-        else:
-            self.request.set_cookie(
-                settings.LANGUAGE_COOKIE_NAME, form.cleaned_data['language']
-            )
-            self.request.set_cookie(
-                settings.TIMEZONE_COOKIE_NAME, form.cleaned_data['timezone']
-            )
-
-        return super().form_valid(form=form)
-
-    def get_object(self):
-        return self.request.user.locale_profile
 
 
 class FaviconRedirectView(RedirectView):