+++ mayan/apps/locales/icons.py
@@ -1,0 +1,9 @@
+from mayan.apps.appearance.classes import Icon
+
+icon_current_user_locale_profile_details = Icon(
+    driver_name='fontawesome', symbol='globe'
+)
+icon_current_user_locale_profile_edit = Icon(
+    driver_name='fontawesome-dual', primary_symbol='globe',
+    secondary_symbol='pencil-alt'
+)