+++ mayan/settings/base.py
@@ -83,6 +83,7 @@ INSTALLED_APPS = (
     'mayan.apps.dynamic_search',
     'mayan.apps.events',
     'mayan.apps.file_caching',
+    'mayan.apps.locales',
     'mayan.apps.lock_manager',
     'mayan.apps.logging',
     'mayan.apps.mimetype',
@@ -138,7 +139,7 @@ MIDDLEWARE = (
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'django.middleware.locale.LocaleMiddleware',
-    'mayan.apps.common.middleware.timezone.TimezoneMiddleware',
+    'mayan.apps.locales.middleware.timezone.TimezoneMiddleware',
     'stronghold.middleware.LoginRequiredMiddleware',
     'mayan.apps.common.middleware.ajax_redirect.AjaxRedirect',
 )